python/requirements.txt
# Greenspot Grocer ETL Requirements
# Install with: pip install -r requirements.txt

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
mysql-connector-python>=8.0.31

# Data validation and cleaning
python-dateutil>=2.8.2

# Logging and utilities
logging>=0.4.9.6

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation (optional)
sphinx>=5.0.0